+++ apps/newsletters/emails.py
@@ -1,3 +1,5 @@
+from django.urls import reverse
+
 from apps.users.emails import EmailAplus as Email
 from apps.users.models import User
 
@@ -40,3 +42,23 @@ class NewsletterEmail(Email):
 class NewsletterEmailAll(NewsletterEmail):
     def get_receivers(self):
         return User.objects.filter(is_active=True).distinct()
+
+
+class NewsletterSettingsNoticeEmail(Email):
+    """One-time service email informing users that their newsletter opt-in
+    may have been reset by a bug. Sent regardless of the current opt-in."""
+
+    template_name = "a4_candy_newsletters/emails/newsletter_settings_notice"
+
+    def __init__(self, user):
+        self._user = user
+
+    def get_receivers(self):
+        return [self._user]
+
+    def get_context(self):
+        context = super().get_context()
+        context["settings_url"] = self.get_host() + reverse(
+            "account_notification_settings"
+        )
+        return context